pyproject.toml
[project]
name = "ace-skyspark-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Andrew Rodgers", email = "andrew@aceiotsolutions.com" }
]
requires-python = ">=3.13"
dependencies = [
    "ace-skyspark-lib",
    "aceiot-models>=0.3.4",
    "click>=8.3.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.7.4",
]

[project.scripts]
ace-skyspark-cli = "ace_skyspark_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyrefly>=0.39.4",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.3",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
ace-skyspark-lib = { path = "../ace-skyspark-lib" }

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyrefly]
python-version = "3.13"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--color=yes",
]
markers = [
    "unit: Unit tests that don't require external services",
    "integration: Integration tests that require FlightDeck and SkySpark",
    "idempotent: Tests that verify idempotent operations",
    "slow: Tests that may take longer to run",
]
asyncio_mode = "auto"
